Move generalizer to its own module
